Create and delete the schedules files in an organized way

+++ server/api/crud/model_monitoring/model_endpoints.py
@@ -16,6 +16,7 @@ import itertools
 import json
 import typing
 
+import fsspec
 import sqlalchemy.orm
 
 import mlrun.artifacts
@@ -35,14 +36,14 @@ from mlrun.utils import logger
 
 
 class _ModelMonitoringSchedulesFile:
-    EMPTY_CONTENT = json.dumps({})
+    INITIAL_CONTENT = json.dumps({})
 
     def __init__(self, project: str, endpoint_id: str) -> None:
-        self._item = mlrun.datastore.store_manager.object(
-            url=mlrun.model_monitoring.helpers._get_monitoring_schedules_file_path(
-                project=project, endpoint_id=endpoint_id
-            )
+        self._path = mlrun.model_monitoring.helpers._get_monitoring_schedules_file_path(
+            project=project, endpoint_id=endpoint_id
         )
+        self._item = mlrun.datastore.store_manager.object(self._path)
+        self._fs = typing.cast(fsspec.AbstractFileSystem, self._item.store.filesystem)
 
     @classmethod
     def from_model_endpoint(
@@ -54,12 +55,22 @@ class _ModelMonitoringSchedulesFile:
         )
 
     def create(self) -> None:
+        """Create a schedules file with initial content - an empty dictionary"""
         logger.debug("Creating model monitoring schedules file", path=self._item.url)
-        self._item.put(self.EMPTY_CONTENT)
+        self._item.put(self.INITIAL_CONTENT)
 
     def delete(self) -> None:
-        logger.debug("Deleting model monitoring schedules file", path=self._item.url)
-        self._item.delete()
+        """Delete schedules file if it exists"""
+        if self._fs.exists(self._path):
+            logger.debug(
+                "Deleting model monitoring schedules file", path=self._item.url
+            )
+            self._item.delete()
+        else:
+            logger.debug(
+                "Model monitoring schedules file does not exist, nothing to delete",
+                path=self._item.url,
+            )
 
 
 class ModelEndpoints:
@@ -140,7 +151,7 @@ class ModelEndpoints:
             # Create monitoring feature set if monitoring found in model endpoint object
             if (
                 model_endpoint.spec.monitoring_mode
-                == mlrun.common.schemas.model_monitoring.ModelMonitoringMode.enabled.value
+                == mlrun.common.schemas.model_monitoring.ModelMonitoringMode.enabled
             ):
                 monitoring_feature_set = cls.create_monitoring_feature_set(
                     features=features,
@@ -178,8 +189,12 @@ class ModelEndpoints:
         )
         model_endpoint_store.write_model_endpoint(endpoint=model_endpoint.flat_dict())
 
-        # Create model monitoring schedules file in any case
-        _ModelMonitoringSchedulesFile.from_model_endpoint(model_endpoint).create()
+        if (
+            model_endpoint.spec.monitoring_mode
+            == mlrun.common.schemas.model_monitoring.ModelMonitoringMode.enabled
+        ):
+            # Create model monitoring schedules file
+            _ModelMonitoringSchedulesFile.from_model_endpoint(model_endpoint).create()
 
         logger.info("Model endpoint created", endpoint_id=model_endpoint.metadata.uid)
 
@@ -621,6 +636,18 @@ class ModelEndpoints:
             model_monitoring_applications=model_monitoring_applications,
             model_monitoring_access_key=model_monitoring_access_key,
         )
+
+        # Delete model monitoring schedules folder
+        folder = mlrun.model_monitoring.helpers._get_monitoring_schedules_folder_path(
+            project_name
+        )
+        fs = typing.cast(
+            fsspec.AbstractFileSystem,
+            mlrun.datastore.store_manager.object(folder).store.filesystem,
+        )
+        if fs.exists(folder):
+            logger.debug("Deleting model monitoring schedules folder", folder=folder)
+            fs.rm(folder, recursive=True)
 
         logger.debug(
             "Successfully deleted model monitoring endpoints resources",
